+++ bin/projects.dart
@@ -40,9 +40,7 @@ List<Project> findProjects() {
     }
 
     final Engine engine;
-    if (fvmPaths.any(absolutePath.startsWith)) {
-      engine = Engine.fvm;
-    } else if (pubspec.dependencies['flutter'] != null) {
+    if (pubspec.dependencies['flutter'] != null) {
       engine = Engine.flutter;
     } else {
       engine = Engine.dart;
@@ -53,12 +51,17 @@ List<Project> findProjects() {
         .split(Platform.pathSeparator)
         .any((e) => e.length > 1 && e.startsWith('.'));
 
+    final hasBuildRunner = pubspec.devDependencies.containsKey('build_runner');
+    final fvm = fvmPaths.any(absolutePath.startsWith);
+
     final project = Project(
       engine: engine,
       path: path,
       config: config,
       example: example,
       hidden: hidden,
+      hasBuildRunner: hasBuildRunner,
+      fvm: fvm,
     );
 
     projects.add(project);
@@ -68,21 +71,21 @@ List<Project> findProjects() {
 }
 
 extension ProjectExtension on Project {
-  Engine _resolveEngine(Command command) {
-    final isCleanCommand = command.args[0] == 'clean';
+  Engine _resolveEngine(ProjectCommand command) {
+    final commandEngine = command.engine;
     final isTestCoverageCommand = command.args.length >= 2 &&
         command.args[0] == 'test' &&
         command.args[1] == '--coverage';
 
     final Engine newEngine;
     final String? message;
-    if (isCleanCommand && !engine.isFlutter) {
-      newEngine = Engine.flutter;
-      message = 'Overriding engine to "flutter" for "clean" command';
-    } else if (isTestCoverageCommand && !engine.isFlutter) {
+    if (commandEngine != null) {
+      newEngine = commandEngine;
+      message = 'Overriding engine to "${commandEngine.name}" for command';
+    } else if (isTestCoverageCommand && engine != Engine.flutter) {
       newEngine = Engine.flutter;
       message = 'Overriding engine to "flutter" for "test --coverage" command';
-    } else if (engine == Engine.fvm && command.noFvm) {
+    } else if (fvm && command.noFvm) {
       newEngine = Engine.flutter;
       message = 'Project uses FVM, but FVM support is disabled: $path';
     } else {
@@ -102,6 +105,10 @@ extension ProjectExtension on Project {
         command.args[0] == 'pub' &&
         command.args[1] == 'get';
 
+    final isBuildRunner = command.args.length >= 2 &&
+        command.args[0] == 'run' &&
+        command.args[1] == 'build_runner';
+
     final bool skip;
     final String? message;
     if (hidden) {
@@ -114,6 +121,10 @@ extension ProjectExtension on Project {
     } else if (isPubGetInExample) {
       // Skip pub get in example projects since it happens anyways
       message = 'Skipping example project: $path';
+      skip = true;
+    } else if (isBuildRunner && !hasBuildRunner) {
+      // Skip build_runner commands if the project doesn't use build_runner
+      message = 'Skipping project without build_runner: $path';
       skip = true;
     } else {
       message = null;
@@ -138,13 +149,18 @@ extension ProjectExtension on Project {
   }
 
   Project resolveWithCommand(Command command) {
-    final resolvedEngine = _resolveEngine(command);
+    final Engine resolvedEngine;
+    if (command is ProjectCommand) {
+      resolvedEngine = _resolveEngine(command);
+    } else {
+      resolvedEngine = engine;
+    }
     final exclude = _defaultExclude(command) || _explicitExclude(command);
     return copyWith(engine: resolvedEngine, exclude: exclude);
   }
 
   Future<Version?> getFlutterVersionOverride(Command command) async {
-    if (engine != Engine.fvm || command.noFvm) return null;
+    if (!fvm || command.noFvm) return null;
 
     try {
       // TODO: Do this a better way (https://github.com/leoafarias/fvm/issues/710)